Extract query since parsing helpers
