+++ src/CodeIndex/Cli/QueryCommandRunner.SinceParsing.cs
@@ -1,0 +1,48 @@
+using System.Globalization;
+
+namespace CodeIndex.Cli;
+
+public static partial class QueryCommandRunner
+{
+    // Accepted ISO 8601 formats for --since / --sinceフィルタで受け付けるISO 8601書式
+    private static readonly string[] Iso8601Formats =
+    [
+        // date only / 日付のみ
+        "yyyy-MM-dd",
+        // minute precision / 分精度
+        "yyyy-MM-ddTHH:mm",
+        "yyyy-MM-ddTHH:mmZ",
+        "yyyy-MM-ddTHH:mmzzz",
+        // second precision / 秒精度
+        "yyyy-MM-ddTHH:mm:ss",
+        "yyyy-MM-ddTHH:mm:ssZ",
+        "yyyy-MM-ddTHH:mm:sszzz",
+        // fractional seconds (1-7 digits via 'F') / 小数秒（1-7桁、'F'で可変長）
+        "yyyy-MM-ddTHH:mm:ss.FFFFFFFZ",
+        "yyyy-MM-ddTHH:mm:ss.FFFFFFFzzz",
+        "yyyy-MM-ddTHH:mm:ss.FFFFFFF",
+        // round-trip format / ラウンドトリップ書式
+        "o",
+    ];
+
+    /// <summary>
+    /// Parse a --since value using invariant ISO 8601 formats only.
+    /// Rejects ambiguous locale-dependent formats like MM/dd/yyyy.
+    /// Offsetless inputs are treated as UTC so the same `--since 2024-01-01T00:00:00`
+    /// resolves to the same logical UTC moment regardless of the caller's timezone
+    /// (Issue #1545). Append `Z` or an explicit offset (`+09:00`) to opt out.
+    /// ISO 8601形式のみで--since値をパースする。MM/dd/yyyyなどロケール依存の曖昧な形式は拒否する。
+    /// オフセットなしの入力はUTCとして扱い、呼び出し側のタイムゾーンに依らず同じUTC時点になる
+    /// （Issue #1545）。明示的にオフセットを付けたい場合は `Z` または `+09:00` 等を付与する。
+    /// </summary>
+    internal static bool TryParseIso8601Since(string value, out DateTime result)
+    {
+        if (DateTimeOffset.TryParseExact(value, Iso8601Formats, CultureInfo.InvariantCulture, DateTimeStyles.AssumeUniversal, out var dto))
+        {
+            result = dto.UtcDateTime;
+            return true;
+        }
+        result = default;
+        return false;
+    }
+}
